backend/pyproject.toml
[project]
name = "ai-news-backend"
version = "0.1.0"
requires-python = ">=3.11"
dependencies = [
  "fastapi>=0.110",
  "uvicorn[standard]>=0.27",
  "pydantic>=2.6",
  "pydantic-settings>=2.2",
  "sqlalchemy>=2.0",
  "asyncpg>=0.29",
  "alembic>=1.13",
  "passlib[bcrypt]>=1.7",
  "python-jose[cryptography]>=3.3",
  "authlib>=1.3",
  "httpx>=0.27",
  "redis>=5.0",
  "structlog>=24.1",
  "orjson>=3.9",
  "langgraph>=0.0.70",
  "google-genai>=0.3.0"
]

[tool.uvicorn]
factory = false
